third_party/WebKit/LayoutTests/typedcssom/stylevalue-subclasses/numeric-objects/cssMathValue-expected.txt

This is a testharness.js-based test.
FAIL Constructing a CSSMathSum with no arguments throws a SyntaxError assert_throws: function "() => new subclass()" threw object "TypeError: Failed to construct 'CSSMathSum': 1 argument required, but only 0 present." ("TypeError") expected object "SyntaxError" ("SyntaxError")
PASS CSSMathSum can be constructed from a single number CSSUnitValue
PASS CSSMathSum can be constructed from more than one number CSSUnitValues
PASS CSSMathSum.operator is readonly
FAIL Constructing a CSSMathProduct with no arguments throws a SyntaxError assert_throws: function "() => new subclass()" threw object "TypeError: Failed to construct 'CSSMathProduct': 1 argument required, but only 0 present." ("TypeError") expected object "SyntaxError" ("SyntaxError")
PASS CSSMathProduct can be constructed from a single number CSSUnitValue
PASS CSSMathProduct can be constructed from more than one number CSSUnitValues
PASS CSSMathProduct.operator is readonly
FAIL Constructing a CSSMathMin with no arguments throws a SyntaxError assert_throws: function "() => new subclass()" threw object "TypeError: Failed to construct 'CSSMathMin': 1 argument required, but only 0 present." ("TypeError") expected object "SyntaxError" ("SyntaxError")
PASS CSSMathMin can be constructed from a single number CSSUnitValue
PASS CSSMathMin can be constructed from more than one number CSSUnitValues
PASS CSSMathMin.operator is readonly
FAIL Constructing a CSSMathMax with no arguments throws a SyntaxError assert_throws: function "() => new subclass()" threw object "TypeError: Failed to construct 'CSSMathMax': 1 argument required, but only 0 present." ("TypeError") expected object "SyntaxError" ("SyntaxError")
PASS CSSMathMax can be constructed from a single number CSSUnitValue
PASS CSSMathMax can be constructed from more than one number CSSUnitValues
PASS CSSMathMax.operator is readonly
Harness: the test ran to completion.
